+++ workspace/tons-of-tests-cli/CHANGELOG.md
@@ -1,5 +1,16 @@
 # @creative-introvert/tons-of-tests-cli
 
+## 1.2.0
+
+### Minor Changes
+
+- Feat: testCases can be a Stream
+
+### Patch Changes
+
+- Updated dependencies
+  - @creative-introvert/tons-of-tests@1.2.0
+
 ## 1.1.1
 
 ### Patch Changes